+++ packages/dart_pdf_editor/example/lib/l10n/app_localizations_th.dart
@@ -437,4 +437,9 @@ class AppLocalizationsTh extends AppLocalizations {
 
   @override
   String get exFileTypeFonts => 'ฟอนต์';
+
+  @override
+  String exExtractedTitle(String title, int part) {
+    return '$title - ส่วนที่ $part.pdf';
+  }
 }